third_party/blink/web_tests/media/track/track-word-breaking.html

<!DOCTYPE html>
<title>Test that line breaks are forced in captions.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-controls.js"></script>
<video autoplay>
    <track src="captions-webvtt/long-word.vtt" kind="captions" default>
</video>
<script>
async_test(function(t) {
    var singleLineCaptionHeight;
    var video = document.querySelector("video");
    video.oncanplaythrough = t.step_func(function() {
        // Record Single line caption height.
        singleLineCaptionHeight = getCaptionLineHeight();
        video.currentTime = 3;
    });

    video.onseeked = t.step_func_done(function() {
        var multiLineCaptionHeight = getCaptionLineHeight();
        assert_greater_than(multiLineCaptionHeight / singleLineCaptionHeight, 2);
    });

    video.src = "../content/test.ogv";

    function getCaptionLineHeight() {
        var captionElement = textTrackContainerElement(video).firstChild.firstChild;
        return captionElement.getBoundingClientRect().height;
    }
});
</script>
